// packages/contracts/generated/MediaRevision.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Id } from "./Id";

export type MediaRevision = { asset_id: Id, asset_revision_id: Id, project_id: Id, original_name: string, original_sha256: string, canonical_sha256: string, canonical_width: number, canonical_height: number, exif_orientation: number, original_to_canonical: Array<number>, source_group_id: Id, };